requirements.txt
grpcio
grpcio-tools
yfinance
cachetools
mysqlclient